+++ qcc/symbol.h
@@ -15,7 +15,7 @@ typedef struct {
     SymbolName name;
     int declaration_line;
     Type type;
-    uint8_t constant_index;
+    uint16_t constant_index;
 } Symbol;
 
 typedef struct {
